// lib/types.ts
export type UserRole = 'CLIENT' | 'PHOTOGRAPHER' | 'ADMIN';
// ...
export interface Account {
  id: string;
  userId: string;
  type: string;
  provider: string;
  providerAccountId: string;
  refresh_token?: string | null;
  access_token?: string | null;
  expires_at?: number | null;
  token_type?: string | null;
  scope?: string | null;
  id_token?: string | null;
  session_state?: string | null;
  user: User;
}

export interface Session {
  id: string;
  sessionToken: string;
  userId: string;
  expires: Date;
  user: User;
}

export interface User {
  id: string;
  name: string;
  email: string;
  emailVerified?: Date | null;
  password: string;
  userRole: UserRole;
  image: string;
  accounts: Account[];
  sessions: Session[];
  photographer?: Photographer | null;
  photographerId?: string | null;
  client?: Client | null;
  admin?: Admin | null;
  chats: Chat[];
  chatIds: string[];
  messagesSent: Message[];
  messagesReceived: Message[];
  createdAt: Date;
  updatedAt: Date;
}

export interface Photographer {
  id: string;
  user: User;
  userId: string;
  coverPhoto?: string | null;
  bio?: string | null;
  featured: string[];
  category: PhotographerCategory[];
  contactDetails?: ContactDetails | null;
  testimonial: Testimonial[];
  feed: FeedImage[];
  packages: Package[];
  albums: Album[];
  Bookings: Booking[];
  BankDetails?: BankDetails | null;
  Earnings?: Earnings | null;
  Report: Report[];
  createdAt: Date;
  updatedAt: Date;
}

export interface Client {
  id: string;
  user: User;
  userId: string;
  testimonial: Testimonial[];
  Booking: Booking[];
  Report: Report[];
  createdAt: Date;
  updatedAt: Date;
}

export interface Admin {
  id: string;
  user: User;
  userId: string;
  createdAt: Date;
  updatedAt: Date;
}

export interface Chat {
  id: string;
  users: User[];
  userIds: string[];
  messages: Message[];
  messagesIds: string[];
  createdAt: Date;
  updatedAt: Date;
}

export interface Message {
  id: string;
  chat: Chat;
  chatId: string;
  sender: User;
  senderId: string;
  receiver: User;
  receiverId: string;
  message: string;
  attachments: Attachment[];
  createdAt: Date;
  updatedAt: Date;
}

export interface Attachment {
  id: string;
  message: Message;
  messageId: string;
  type: string;
  url: string;
}